PHP Update
Unified State Response
$_SESSION functionality
new functions

+++ delete_page.php
@@ -1,5 +1,6 @@
 <?php
 	include("config.php");
+	try_session();
 	
 	if($_SERVER["REQUEST_METHOD"] == "POST")
 	{
@@ -8,27 +9,33 @@
 		$response = new \stdClass();
 		
 		$pageId = $obj['id'];
-		$access = $obj['access'];
-		
-		if($access > 3)
+		$access = $_SESSION['access'];
+		if($_SESSION['isLoggedIn'])
 		{
-			$conn = getDatabase();
-			
-			$stmt = $conn->prepare("DELETE FROM Sections WHERE page_id = ?");
-			$stmt->bind_param("i", $pageId);
-			$stmt->execute();
-			$stmt->close();
-			
-			$stmtDeletePage = $conn->prepare("DELETE FROM Pages WHERE id = ?");
-			$stmtDeletePage->bind_param("i", $pageId);
-			$stmtDeletePage->execute();
-			$stmtDeletePage->close();
-			
-			$response->status = "Success";
+			if($access > 3)
+			{
+				$conn = getDatabase();
+				
+				$stmt = $conn->prepare("DELETE FROM Sections WHERE page_id = ?");
+				$stmt->bind_param("i", $pageId);
+				$stmt->execute();
+				$stmt->close();
+				
+				$stmtDeletePage = $conn->prepare("DELETE FROM Pages WHERE id = ?");
+				$stmtDeletePage->bind_param("i", $pageId);
+				$stmtDeletePage->execute();
+				$stmtDeletePage->close();
+				
+				$response->state = 1; //Success
+			}
+			else
+			{
+				$response->state = 4 //Not logged in, access denied;
+			}
 		}
 		else
 		{
-			$response->status = "Access Denied";
+			$response->state = 2; //User not logged in
 		}
 		
 		echo json_encode($response);
